+++ nevergrad/parametrization/data.py
@@ -8,6 +8,7 @@ import warnings
 import numpy as np
 import nevergrad.common.typing as tp
 from . import core
+from . import layers
 from .container import Dict
 from . import utils
 from . import transforms as trans
@@ -339,7 +340,8 @@ class Data(core.Parameter):
         difficult. It is especially ill-advised to use this with a range smaller than 10, or
         a sigma lower than 1. In those cases, you should rather use a TransitionChoice instead.
         """
-        self.integer = True
+        self.add_layer(layers.IntegerCasting())
+        # self.integer = True
         return self
 
     # pylint: disable=unused-argument
